Moved Camera enums into a standalone extension file
