+++ Sources/UI/Camera/Camera+PovioKit.swift
@@ -1,0 +1,57 @@
+//
+//  Camera+PovioKit.swift
+//  PovioKit
+//
+//  Created by Ndriqim Nagavci on 21/10/2022.
+//  Copyright © 2022 Povio Inc. All rights reserved.
+//
+
+import AVFoundation
+
+public extension Camera {
+  enum CameraPosition {
+    case back
+    case front
+  }
+
+  enum CameraAuthorizationStatus {
+    case authorized
+    case denied
+    case notDetermined
+  }
+
+  enum MediaType {
+    case video
+    case audio
+  }
+  
+  enum Error: Swift.Error {
+    case unavailable
+    case missingSession
+    case missingInput
+    case missingOutput
+    case invalidImage
+  }
+}
+
+extension Camera.CameraPosition {
+  var asAVCaptureDevicePosition: AVCaptureDevice.Position {
+    switch self {
+    case .back:
+      return .back
+    case .front:
+      return .front
+    }
+  }
+}
+
+extension Camera.MediaType {
+  var asAVMediaType: AVMediaType {
+    switch self {
+    case .video:
+      return .video
+    case .audio:
+      return .audio
+    }
+  }
+}
